feat: Complete ShazaPiano implementation
- ✅ All 6 pages implemented (Home, Previews, Player, Practice, Settings, History)
- ✅ All widgets (BigRecordButton, ModeChip, VideoTile, AppLogo, PaywallModal)
- ✅ Full navigation flow according to PDFs
- ✅ Auto-stop recording after 8 seconds
- ✅ Backend integration (upload, process, display)
- ✅ Firebase complete (auth, firestore, analytics, crashlytics)
- ✅ IAP flow (purchase, restore)
- ✅ Practice mode with pitch detection (MPM algorithm)
- ✅ Responsive layout (portrait + landscape)
- ✅ Overflow bugs fixed
- ✅ App icon updated (turquoise piano with treble clef)
- ✅ Security: google-services.json excluded from git
- 🎹 App fully functional on device!

+++ OneDrive/Bureau/shazam piano/app/lib/presentation/pages/practice/pitch_detector.dart
@@ -140,7 +140,7 @@ class PitchDetector {
 
   /// Convert MIDI note to frequency
   double midiNoteToFrequency(int note) {
-    return 440 * pow(2, (note - 69) / 12);
+    return (440 * pow(2, (note - 69) / 12)).toDouble();
   }
 
   /// Calculate cents difference between two frequencies
@@ -168,4 +168,5 @@ enum NoteAccuracy {
   wrong,    // >50 cents
   miss,     // No note detected
 }
+
